docs: add TypeDoc annotations
