+++ package.json
@@ -44,6 +44,7 @@
     "mocha": "^10.2.0",
     "nyc": "^15.1.0",
     "ts-node": "^10.9.1",
+    "typedoc": "^0.25.3",
     "typescript": "^5.2.2"
   },
   "scripts": {
