+++ src/codeActionHandler.ts
@@ -9,6 +9,12 @@ import { codeActionsMap } from './diagnosticsHandler';
 import { globalConfig } from './config';
 import { RHDA_DIAGNOSTIC_SOURCE } from './constants';
 
+/**
+ * Retrieves code actions based on diagnostics and file type.
+ * @param diagnostics - An array of available diagnostics.
+ * @param fileType - The type of the file based on ecosystem (e.g., 'pom.xml').
+ * @returns An array of CodeAction objects to be made available to the user.
+ */
 function getDiagnosticsCodeActions( diagnostics: Diagnostic[], fileType: string ): CodeAction[] {
     const codeActions: CodeAction[] = [];
     let hasRhdaDiagonostic: boolean = false;
@@ -18,7 +24,7 @@ function getDiagnosticsCodeActions( diagnostics: Diagnostic[], fileType: string 
         if (codeAction) {
             
             if (fileType === 'pom.xml') {
-                // add RedHat repository recommendation command to action
+                // add Red Hat repository recommendation command to action
                 codeAction.command = {
                 title: 'RedHat repository recommendation',
                 command: globalConfig.triggerRHRepositoryRecommendationNotification,
@@ -38,6 +44,10 @@ function getDiagnosticsCodeActions( diagnostics: Diagnostic[], fileType: string 
     return codeActions;
 }
 
+/**
+ * Generates a code action for a detailed RHDA report on the analyzed manifest file.
+ * @returns A CodeAction object for an RHDA report.
+ */
 function generateFullStackAnalysisAction(): CodeAction {
     return {
         title: 'Detailed Vulnerability Report',
