+++ src/providers/go.mod.ts
@@ -1,9 +1,28 @@
+/* --------------------------------------------------------------------------------------------
+ * Copyright (c) Red Hat
+ * Licensed under the Apache-2.0 License. See License.txt in the project root for license information.
+ * ------------------------------------------------------------------------------------------ */
 'use strict';
 
 import { IDependencyProvider, EcosystemDependencyResolver, IDependency, Dependency } from '../collector';
-import { semVerRegExp } from '../utils'
 
-/* Process entries found in the go.mod file */
+/* Please note :: There is an issue with the usage of semverRegex Node.js package in this code.
+ * Often times it fails to recognize versions that contain an added suffix, usually including extra details such as a timestamp and a commit hash.
+ * At the moment, using regex directly to extract versions inclusively. */
+
+/**
+ * Executes a regular expression pattern match for Semantic Versioning (SemVer) within a given string.
+ * @param str - The string to search for a Semantic Versioning pattern.
+ * @returns An array of matched results for the Semantic Versioning pattern.
+ */
+export function semVerRegExp(str: string): RegExpExecArray {
+    const regExp = /(?<=^v?|\sv?)(?:0|[1-9]\d*)\.(?:0|[1-9]\d*)\.(?:0|[1-9]\d*)(?:-(?:0|[1-9]\d*|[\da-z-]*[a-z-][\da-z-]*)(?:\.(?:0|[1-9]\d*|[\da-z-]*[a-z-][\da-z-]*))*)?(?:\+[\da-z-]+(?:\.[\da-z-]+)*)?(?=$|\s)/ig;
+    return regExp.exec(str);
+  }
+
+/**
+ * Process entries found in the go.mod file.
+ */
 export class DependencyProvider extends EcosystemDependencyResolver implements IDependencyProvider {
     replacementMap: Map<string, IDependency> = new Map<string, IDependency>();
     
@@ -11,35 +30,56 @@ export class DependencyProvider extends EcosystemDependencyResolver implements I
         super('golang'); // set ecosystem to 'golang'
     }
 
+    /**
+     * Parses the provided string as an array of lines.
+     * @param contents - The string content to parse into lines.
+     * @returns An array of strings representing lines from the provided content.
+     */
     static parseTxtDoc(contents: string): string[] {
         return contents.split('\n');
     }
 
+    /**
+     * Cleans the given string by removing specific characters and words, 
+     * such as 'require' and 'replace', parentheses, and consecutive spaces from the provided string.
+     * Additionally, trims any leading or trailing whitespace.
+     * @param line - The string to be cleaned.
+     * @returns The cleaned string.
+     */
     static clean(line: string): string {
         return line.replace(/require|replace|\(|\)/g, '')
         .replace(/\s+/g, ' ')
         .trim();
     }
     
+    /**
+     * Extracts dependency data from the provided line.
+     * @param line - The line to extract dependency data from.
+     * @returns An object containing dependency data, or null if no matching version is found.
+     */
     static getDependencyData(line: string): { name: string, version: string, index: number } | null  {
         const versionMatches: RegExpExecArray = semVerRegExp(line);
         if (versionMatches && versionMatches.length > 0) {
             const depName = DependencyProvider.clean(line).split(' ')[0];
-            return {name: depName, version: versionMatches[0], index: versionMatches.index}
+            return {name: depName, version: versionMatches[0], index: versionMatches.index};
         }
         return null;
     }
 
+    /**
+     * Registers a replacement dependency in the replacement map.
+     * @param line - The line containing the replacement statement.
+     * @param index - The index of the line in the file.
+     */
     private registerReplacement(line: string, index: number) {
-        // split the replace statements by '=>'
         const lineData: string[] = line.split('=>');
-        if (lineData.length !== 2) return;
+        if (lineData.length !== 2) { return; }
 
         let originalDepData = DependencyProvider.getDependencyData(lineData[0]);
         const replacementDepData = DependencyProvider.getDependencyData(lineData[1]);
 
-        if (!originalDepData) originalDepData = {name: DependencyProvider.clean(lineData[0]), version: null, index: null}
-        if (!replacementDepData) return;
+        if (!originalDepData) { originalDepData = {name: DependencyProvider.clean(lineData[0]), version: null, index: null}; }
+        if (!replacementDepData) { return; }
 
         const replaceDependency = new Dependency(
             { value: replacementDepData.name, position: { line: 0, column: 0 } },
@@ -49,10 +89,16 @@ export class DependencyProvider extends EcosystemDependencyResolver implements I
         this.replacementMap.set(originalDepData.name + (originalDepData.version ? ('@v' + originalDepData.version) : ''), replaceDependency); 
     }
 
+    /**
+     * Parses a line from the file and extracts dependency information.
+     * @param line - The line to parse for dependency information.
+     * @param index - The index of the line in the file.
+     * @returns An IDependency object representing the parsed dependency or null if no dependency is found.
+     */
     private parseLine(line: string, index: number): IDependency | null {
 
         line = line.split('//')[0]; // Remove comments
-        if (!DependencyProvider.clean(line)) return null; // Skip lines without dependencies
+        if (!DependencyProvider.clean(line)) { return null; } // Skip lines without dependencies
 
         if (line.includes('=>')) {
             // stash replacement dependencies for replacement
@@ -61,7 +107,7 @@ export class DependencyProvider extends EcosystemDependencyResolver implements I
         }
 
         const depData = DependencyProvider.getDependencyData(line);
-        if (!depData) return null;
+        if (!depData) { return null; }
 
         return new Dependency(
             { value: depData.name, position: { line: 0, column: 0 } },
@@ -69,10 +115,20 @@ export class DependencyProvider extends EcosystemDependencyResolver implements I
         );
     }
 
+    /**
+     * Applies replacement dependency from replacement map to the provided dependency.
+     * @param dep - The dependency to be checked and replaced if necessary.
+     * @returns The replaced dependency or the original one if no replacement is found.
+     */
     private  applyReplaceMap(dep: IDependency): IDependency {
         return this.replacementMap.get(dep.name.value + '@' + dep.version.value) || this.replacementMap.get(dep.name.value) || dep;
     }
 
+    /**
+     * Extracts dependencies from lines parsed from the file.
+     * @param lines - An array of strings representing lines from the file.
+     * @returns An array of IDependency objects representing extracted dependencies.
+     */
     private extractDependenciesFromLines(lines :string[]): IDependency[] {
         let isExcluded: boolean = false;
         const goModDeps: IDependency[] = lines.reduce((dependencies: IDependency[], line: string, index: number) => {
@@ -91,7 +147,6 @@ export class DependencyProvider extends EcosystemDependencyResolver implements I
                 return dependencies;
             }
 
-            // parse included lines for dependencies
             const parsedDependency: IDependency = this.parseLine(line, index);
             if (parsedDependency) {
                 dependencies.push(parsedDependency);
@@ -101,10 +156,14 @@ export class DependencyProvider extends EcosystemDependencyResolver implements I
 
         }, []);
 
-        // apply replacement dependencies
         return goModDeps.map(goModDep => this.applyReplaceMap(goModDep));
     }
 
+    /**
+     * Collects dependencies from the provided manifest contents.
+     * @param contents - The manifest content to collect dependencies from.
+     * @returns A Promise resolving to an array of IDependency objects representing collected dependencies.
+     */
     async collect(contents: string): Promise<IDependency[]> {
         const lines: string[] = DependencyProvider.parseTxtDoc(contents);
         return this.extractDependenciesFromLines(lines);
