+++ src/providers/package.json.ts
@@ -1,8 +1,15 @@
+/* --------------------------------------------------------------------------------------------
+ * Copyright (c) Red Hat
+ * Licensed under the Apache-2.0 License. See License.txt in the project root for license information.
+ * ------------------------------------------------------------------------------------------ */
 'use strict';
 
 import jsonAst from 'json-to-ast';
 import { IDependencyProvider, EcosystemDependencyResolver, IDependency, Dependency } from '../collector';
 
+/**
+ * Process entries found in the package.json file.
+ */
 export class DependencyProvider extends EcosystemDependencyResolver implements IDependencyProvider {
     private classes: string[] = ['dependencies'];
 
@@ -10,12 +17,22 @@ export class DependencyProvider extends EcosystemDependencyResolver implements I
         super('npm'); // set ecosystem to 'npm'
     }
 
+    /**
+     * Parses the provided manifest content into a JSON AST.
+     * @param contents - The manifest content to parse.
+     * @returns The parsed JSON AST.
+     */
     private parseJson(contents: string): jsonAst {
         return jsonAst(contents || '{}');
     }
 
-    private mapDependencies(jsonAst: jsonAst): IDependency[] {
-        return jsonAst.children
+    /**
+     * Maps dependencies from the parsed JSON AST to IDependency objects.
+     * @param jsonAst - The parsed JSON AST to map dependencies from.
+     * @returns An array of IDependency objects representing the dependencies.
+     */
+    private mapDependencies(ast: jsonAst): IDependency[] {
+        return ast.children
                 .filter(c => this.classes.includes(c.key.value))
                 .flatMap(c => c.value.children)
                 .map(c => {
@@ -26,10 +43,15 @@ export class DependencyProvider extends EcosystemDependencyResolver implements I
                 });
     }
 
+    /**
+     * Collects dependencies from the provided manifest contents.
+     * @param contents - The manifest content to collect dependencies from.
+     * @returns A Promise resolving to an array of IDependency objects representing collected dependencies.
+     */
     async collect(contents: string): Promise<IDependency[]> {
         try {
-            const jsonAst: jsonAst = this.parseJson(contents);
-            return this.mapDependencies(jsonAst);
+            const ast: jsonAst = this.parseJson(contents);
+            return this.mapDependencies(ast);
         } catch (err) {
             if (err instanceof SyntaxError) {
                 return [];
