+++ src/server.ts
@@ -13,15 +13,25 @@ import { globalConfig } from './config';
 import { AnalysisLSPServer } from './fileHandler';
 import { getDiagnosticsCodeActions } from './codeActionHandler';
 
-// declare timeout identifier to track delays for server.handleFileEvent execution
+/**
+ * Declares timeout identifier to track delays for server.handleFileEvent execution
+ */
 let checkDelay: NodeJS.Timeout;
 
-// Create a connection for the server, using Node's IPC as a transport.
+/**
+ * Represents the connection used for the server, using Node's IPC as a transport.
+ */
 const connection: Connection = createConnection(ProposedFeatures.all);
+
+/**
+ * Represents the documents managed by the server.
+ */
 const documents: TextDocuments<TextDocument> = new TextDocuments(TextDocument);
 documents.listen(connection);
 
-// Sets up the connection's initialization event handler.
+/**
+ * Sets up the connection's initialization event handler.
+ */
 let hasConfigurationCapability: boolean = false;
 connection.onInitialize((params): InitializeResult => {
     const capabilities = params.capabilities;
@@ -36,30 +46,39 @@ connection.onInitialize((params): InitializeResult => {
     };
 });
 
-// Registers a callback when the connection is fully initialized.
+/**
+ * Registers a callback when the connection is fully initialized.
+ */
 connection.onInitialized(() => {
     if (hasConfigurationCapability) {
-        // Register for all configuration changes.
         connection.client.register(DidChangeConfigurationNotification.type, undefined);
     }
 });
 
+/**
+ * Represents the server handling the Language Server Protocol requests and notifications.
+ */
 const server = new AnalysisLSPServer(connection);
 
-// triggered when document is opened
+/**
+ * On open document event handler
+ */
 connection.onDidOpenTextDocument((params) => {
     server.handleFileEvent(params.textDocument.uri, params.textDocument.text);
 });
 
-// triggered when document is saved
+/**
+ * On save document event handler
+ */
 connection.onDidSaveTextDocument((params) => {
     clearTimeout(checkDelay);
     server.handleFileEvent(params.textDocument.uri, server.files.fileData[params.textDocument.uri]);
 });
 
-// triggered when changes have been applied to document
+/**
+ * On changes applied to document event handler
+ */
 connection.onDidChangeTextDocument((params) => {
-    /* Update internal state for code lenses */
     server.files.fileData[params.textDocument.uri] = params.contentChanges[0].text;
     clearTimeout(checkDelay);
     checkDelay = setTimeout(() => {
@@ -67,23 +86,28 @@ connection.onDidChangeTextDocument((params) => {
     }, 3000);
 });
 
-// triggered when document is closed
+/**
+ * On close document event handler
+ */
 connection.onDidCloseTextDocument(() => {
     clearTimeout(checkDelay);
 });
 
-// Registering a callback when the configuration changes.
+/**
+ * Registers a callback when the configuration changes.
+ */
 connection.onDidChangeConfiguration(() => {
     if (hasConfigurationCapability) {
-        // Fetching the workspace configuration from the client.
         server.conn.workspace.getConfiguration()
         .then((data) => {
-            // Updating global settings based on the fetched configuration data.
             globalConfig.updateConfig(data);
         });
     }
 });
 
+/**
+ * Handles code action requests from client.
+ */
 connection.onCodeAction((params): CodeAction[] => {
     return getDiagnosticsCodeActions(params.context.diagnostics, path.basename(params.textDocument.uri));
 });
