+++ src/vulnerability.ts
@@ -10,8 +10,17 @@ import { IDependencyProvider } from './collector';
 import { DependencyData } from './componentAnalysis';
 import { RHDA_DIAGNOSTIC_SOURCE } from './constants';
 
-/* Vulnerability data along with package name and version */
+/**
+ * Stores vulnerability data of a specific dependency.
+ */
 class Vulnerability {
+    /**
+     * Creates a new instance of Vulnerability.
+     * @param provider - The dependency provider of the corresponding ecosystem.
+     * @param range - The text range within the document.
+     * @param ref - The reference name of the dependency.
+     * @param dependencyData - All vulnerability data regarding the dependency.
+     */
     constructor(
         private provider: IDependencyProvider,
         private range: Range,
@@ -19,6 +28,10 @@ class Vulnerability {
         private dependencyData: DependencyData[],
       ) {}
 
+    /**
+     * Creates a diagnostic object based on vulnerability data.
+     * @returns A Diagnostic object representing the vulnerability.
+     */
     getDiagnostic(): Diagnostic {
 
         const diagSeverity = DiagnosticSeverity.Error;
